+++ Setup/Patch/Data/MigrateProductAttributesPatch.php
@@ -9,8 +9,6 @@ namespace Dhl\ShippingCore\Setup\Patch\Data;
 use Dhl\ShippingCore\Model\Attribute\Migrate;
 use Magento\Catalog\Model\Product;
 use Magento\Eav\Setup\EavSetup;
-use Magento\Eav\Setup\EavSetupFactory;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
 use Magento\Framework\Setup\Patch\DataPatchInterface;
 use Magento\Framework\Setup\Patch\PatchVersionInterface;